refactor: divide alarm screen components into top section and bottom section
